Fix bug after merge

+++ frontend/src/components/feed/item/Writer.vue
@@ -20,8 +20,12 @@
       class="text-decoration-none"
     >
       <v-list-item-content>
-        <v-list-item-title class="text-left red--text text--lighten-2">{{ user.unick }}</v-list-item-title>
-        <v-list-item-subtitle class="text-left red--text text--lighten-2">{{ user.uid }}</v-list-item-subtitle>
+        <v-list-item-title class="text-left red--text text--lighten-2">{{
+          user.unick
+        }}</v-list-item-title>
+        <v-list-item-subtitle class="text-left red--text text--lighten-2">{{
+          user.uid
+        }}</v-list-item-subtitle>
       </v-list-item-content>
     </router-link>
     <v-spacer></v-spacer>
@@ -33,17 +37,28 @@
       </template>
 
       <v-list class="text-center">
-        <v-list-item v-if="user.unick === authUserUnick" @click="moveToUpdateFeed">
-          <v-list-item-title class="blue--text text-lighten-2">게시글 수정</v-list-item-title>
+        <v-list-item
+          v-if="user.unick === authUserUnick"
+          @click="moveToUpdateFeed"
+        >
+          <v-list-item-title class="blue--text text-lighten-2"
+            >게시글 수정</v-list-item-title
+          >
         </v-list-item>
         <v-list-item v-if="user.unick === authUserUnick" @click="deleteFeedNow">
-          <v-list-item-title class="red--text text-lighten-2">게시글 삭제</v-list-item-title>
+          <v-list-item-title class="red--text text-lighten-2"
+            >게시글 삭제</v-list-item-title
+          >
         </v-list-item>
         <v-list-item @click="() => {}">
-          <v-list-item-title class="red--text text-lighten-2">게시글 신고</v-list-item-title>
+          <v-list-item-title class="red--text text-lighten-2"
+            >게시글 신고</v-list-item-title
+          >
         </v-list-item>
         <v-list-item @click="() => {}">
-          <v-list-item-title class="blue--text text-lighten-2">취소</v-list-item-title>
+          <v-list-item-title class="blue--text text-lighten-2"
+            >취소</v-list-item-title
+          >
         </v-list-item>
       </v-list>
     </v-menu>
@@ -90,8 +105,10 @@ export default {
         icon: "warning",
         buttons: true,
         dangerMode: true,
-      }).then(() => {
-        this.deleteFeed(this.$route.params.fid);
+      }).then((doDelete) => {
+        if (doDelete) {
+          this.deleteFeed(this.$route.params.fid);
+        }
       });
     },
   },
