+++ frontend/src/views/feed/CommentView.vue
@@ -3,18 +3,12 @@
     <v-container class="mb-12">
       <v-row class="text-center" align="center" justify="center">
         <v-col cols="12">
-          <v-card flat class="text-left mx-auto" max-width="614" v-if="selectedFeed">
-            <p>
-              <strong>{{ selectedFeed.feed.title }}</strong>
-            </p>
-            <p>{{ selectedFeed.feed.content }}</p>
-          </v-card>
-          <v-spacer>
-            <br />
-          </v-spacer>
           <v-card flat class="mx-auto myCard" max-width="614">
             <v-row class="ma-0">
-              <v-list-item-avatar class="ma-auto mr-2" :color="authUserImgRoute ? 'white' : 'grey'">
+              <v-list-item-avatar
+                class="ma-auto mr-2"
+                :color="authUserImgRoute ? 'white' : 'grey'"
+              >
                 <v-icon v-if="!authUserImgRoute" dark>mdi-account</v-icon>
                 <v-img v-if="authUserImgRoute" :src="authUserImgRoute" />
               </v-list-item-avatar>
@@ -52,17 +46,28 @@
                 </template>-->
               </v-text-field>
             </v-row>
-            <div v-for="(comment,idx) in comments.comments" :key="comment.comment.id">
-              <v-list-item class="ma-0 pa-0 align-start" :id="'a'+comment.comment.id">
+            <div
+              v-for="(comment, idx) in comments.comments"
+              :key="comment.comment.id"
+            >
+              <v-list-item
+                class="ma-0 pa-0 align-start"
+                :id="'a' + comment.comment.id"
+              >
                 <router-link
-                  :to="{ name: 'UserDetail', params: { uid: comment.comment.user.uid } }"
+                  :to="{
+                    name: 'UserDetail',
+                    params: { uid: comment.comment.user.uid },
+                  }"
                   class="text-decoration-none"
                 >
                   <v-list-item-avatar
                     class="mr-2"
                     :color="comment.comment.user.uprofile ? 'white' : 'grey'"
                   >
-                    <v-icon v-if="!comment.comment.user.uprofile" dark>mdi-account</v-icon>
+                    <v-icon v-if="!comment.comment.user.uprofile" dark
+                      >mdi-account</v-icon
+                    >
                     <v-img
                       v-if="comment.comment.user.uprofile"
                       :src="media_dir + comment.comment.user.uprofile.name"
@@ -76,24 +81,33 @@
                   >-->
                   <v-list-item-title class="mb-1 black--text">
                     {{ comment.comment.user.unick }}
-                    <span
-                      style="font-size: 0.7rem; color: grey;"
-                    >{{ computeYMD(comment.comment.regdate) }}</span>
+                    <span style="font-size: 0.7rem; color: grey;">{{
+                      computeYMD(comment.comment.regdate)
+                    }}</span>
                   </v-list-item-title>
                   <!-- </router-link> -->
-                  <v-list-item-subtitle class="black--text mb-1" style="white-space:normal">
+                  <v-list-item-subtitle
+                    class="black--text mb-1"
+                    style="white-space:normal"
+                  >
                     {{ comment.comment.content }}
-                    <span
-                      style="font-size: 0.7rem; color: grey;"
-                    >{{ comment.comment.editdate ? "(수정됨)" : "" }}</span>
+                    <span style="font-size: 0.7rem; color: grey;">{{
+                      comment.comment.editdate ? "(수정됨)" : ""
+                    }}</span>
                   </v-list-item-subtitle>
-                  <v-list-item-subtitle class="gray--text" style="font-size: 0.7rem">
+                  <v-list-item-subtitle
+                    class="gray--text"
+                    style="font-size: 0.7rem"
+                  >
                     <!-- {{ computeYMD(comment.comment.regdate) }} -->
                     <!-- {{ comment.comment.editdate ? "(수정됨)" : "" }} -->
                     좋아요 {{ comment.likeCount }}개
                   </v-list-item-subtitle>
                 </v-list-item-content>
-                <v-menu v-if="authUserUnick === comment.comment.user.unick" bottom>
+                <v-menu
+                  v-if="authUserUnick === comment.comment.user.unick"
+                  bottom
+                >
                   <template v-slot:activator="{ on, attrs }">
                     <v-btn color="#ff6666" icon v-bind="attrs" v-on="on">
                       <v-icon>mdi-dots-horizontal</v-icon>
@@ -106,19 +120,27 @@
                       v-if="authUserUnick === comment.comment.user.unick"
                       @click="showEdit(comment.comment.id)"
                     >
-                      <v-list-item-title class="blue--text text-lighten-2">댓글 수정</v-list-item-title>
+                      <v-list-item-title class="blue--text text-lighten-2"
+                        >댓글 수정</v-list-item-title
+                      >
                     </v-list-item>
                     <v-list-item
                       v-if="authUserUnick === comment.comment.user.unick"
                       @click="deleteCommentAndFetch(comment.comment.id)"
                     >
-                      <v-list-item-title class="red--text text-lighten-2">댓글 삭제</v-list-item-title>
+                      <v-list-item-title class="red--text text-lighten-2"
+                        >댓글 삭제</v-list-item-title
+                      >
                     </v-list-item>
                     <v-list-item @click="() => {}">
-                      <v-list-item-title class="red--text text-lighten-2">댓글 신고</v-list-item-title>
+                      <v-list-item-title class="red--text text-lighten-2"
+                        >댓글 신고</v-list-item-title
+                      >
                     </v-list-item>
                     <v-list-item @click="() => {}">
-                      <v-list-item-title class="blue--text text-lighten-2">취소</v-list-item-title>
+                      <v-list-item-title class="blue--text text-lighten-2"
+                        >취소</v-list-item-title
+                      >
                     </v-list-item>
                   </v-list>
                 </v-menu>
@@ -128,10 +150,12 @@
                   color="#ff6666"
                   icon
                   x-small
-                  @click="commentLike({
-                  'comment': comment,
-                  'idx': idx
-                })"
+                  @click="
+                    commentLike({
+                      comment: comment,
+                      idx: idx,
+                    })
+                  "
                 >
                   <v-icon v-if="comment.islike">mdi-heart</v-icon>
                   <v-icon v-if="!comment.islike">mdi-heart-outline</v-icon>
@@ -140,7 +164,7 @@
               </v-list-item>
               <EditComment
                 class="update"
-                :id="'b'+comment.comment.id"
+                :id="'b' + comment.comment.id"
                 :cid="comment.comment.id"
                 style="display: none;"
                 :comment="comment.comment"
@@ -152,7 +176,16 @@
         </v-col>
       </v-row>
     </v-container>
-    <v-btn color="#ff6666" elevation="24" fixed bottom left fab @click="back()" class="mb-14">
+    <v-btn
+      color="#ff6666"
+      elevation="24"
+      fixed
+      bottom
+      left
+      fab
+      @click="back()"
+      class="mb-14"
+    >
       <v-icon color="#ffffff">mdi-arrow-left-bold</v-icon>
     </v-btn>
   </div>
@@ -215,22 +248,24 @@ export default {
         icon: "warning",
         buttons: true,
         dangerMode: true,
-      }).then(() => {
-        this.deleteComment(id)
-          .then(() => {
-            this.fetchComments(this.fid);
-          })
-          .then(() => {
-            this.comments = this.$store.state.comments;
-          })
-          .catch((err) => console.log(err.response));
+      }).then((doDelete) => {
+        if (doDelete) {
+          this.deleteComment(id)
+            .then(() => {
+              this.fetchComments(this.fid);
+            })
+            .then(() => {
+              this.comments = this.$store.state.comments;
+            })
+            .catch((err) => console.log(err.response));
+        }
       });
     },
     computeYMD(regdate) {
       var ymd =
         parseInt(new Date().getTime() / 1000) -
         parseInt(new Date(regdate).getTime() / 1000);
-      var ymd2 = function (ymd) {
+      var ymd2 = function(ymd) {
         if (ymd < 60) {
           return `${ymd}초 전`;
         } else if (ymd < 3600) {
